upgraded transport report: add completed delivery trip service

+++ src/main/java/com/hust/baseweb/applications/tms/document/aggregation/TransportCustomer.java
@@ -20,10 +20,10 @@ public class TransportCustomer extends TransportReport<TransportCustomer.Id> {
     private Id id;
 
     public TransportCustomer(Id id,
-                             Double cost,
-                             Double totalDistance,
-                             Double numberTrips,
-                             Double totalWeight) {
+                             Long cost,
+                             Integer totalDistance,
+                             Integer numberTrips,
+                             Integer totalWeight) {
         super(cost, totalDistance, numberTrips, totalWeight);
         this.id = id;
     }
